requirements.txt
numpy
opencv2
pillow
gradio
ultralytics
rapidocr
matplotlib
